requirements.txt
openai
langchain
tiktoken
jupyterlab
python-dotenv
